[addons] use CAddonInfo everywhere and prevent CBinaryAddonBase
Majorly is the handling changed, as CBinaryAddonBase becomes
created on background and register his now in binary addon manager.
On last destruct it becomes removed from there.

This thought to fix problems with to late or missing report on binary addon
manager and remove obsolete parts (them was mostly thought before with cpluff).