[video] fix "Scan to library" calls about archives
Before it has not worked as it expected a video file in given directory.
By archives was there the archive name instead of video files.

This checks now the used folder about entries, if now only **1** and them a **directory only**
becomes it opened and checked again. This done so long until a path is found
where have more as 1 entry or no more is a directory.

E.g. this has brought problems:
- rar://some_path_to_archive/the_movie_name/the_archive.rar/the_movie_name.avi
where scan is started on
- file://some_path_to_archive/the_movie_name
The name there was needed to have for search

e.g there the "the_archive.rar" is only one entry, seen as directory and why added the check about.

This then also work with:
- rar://some_path_to_archive/the_movie_name/the_archive.rar/some_unusable_or_usable_name/the_movie_name.avi