[addons][vfs] not allow disable, uninstall and update in case of use
Before was there easily possible to uninstall the addon if it was in use.
There becomes a check added and the CVFSAddonCache asked about usage of
addon, if yes the parts becomes prevented.

To make this, is a helper class added where register on CAddonDll. As the
binary addons use instance handlers, becomes it hard to make another way.
A instance class is for started place, mostly independent from CAddon
and as there can be not easy to maintain.

By VFS is a bit different instance use, there is always only one, where
handle itself open of different parts at same time.
This checks now how much his shared_ptr is in use, if higher as 1 report
that addon class outside used.

But the added CAddonDll parts can be also used for multiple instances within
addon, only need then his related manager arranged to them.