[æddons][gui] move GUIDialogAddonSettings.[cpp|h] to new addon gui folder
This done to have also in gui folder where was add on previous commit.
Somewhere in future and also wanted since long time should be the GUI
indepenedent from other parts.

Before was this in settings folder, but as this is a GUI is his better
place in new gui folder.